refactor : 구조를 개선한다.
